+++ organize/filters/__init__.py
@@ -2,4 +2,5 @@ from .created import Created
 from .extension import Extension
 from .filename import Filename
 from .last_modified import LastModified
+from .python import Python
 from .regex import Regex